{flow、functioncall}: refactor flow and function call response processor ​​to better define its functional boundarie.
